requirements.txt
sentence-transformers==2.2.2
huggingface-hub==0.16.4
transformers==4.30.2
torch==2.2.0
pinecone-client==2.2.4
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
feedparser==6.0.10
beautifulsoup4==4.12.3
google-generativeai==0.3.2
langchain==0.1.9
langchain-google-genai==0.0.9
streamlit==1.31.1
python-crontab==3.0.0 